+++ dotfiles/.config/nwg-dock-hyprland/launch.sh
@@ -17,6 +17,7 @@ if [ ! -f $HOME/.config/ml4w/settings/dock-disabled ]; then
     # nwg-dock-hyprland -i 90 -p bottom -ico awn-applet -mb 5 -d -s style.css -c "/home/unicorn/.config/eww/scripts/applaucher.sh"
     if [ -f $HOME/.config/ml4w/settings/dock-autohide ]; then
         nwg-dock-hyprland -d -i 90 -w 5 -mb 2 -x -c "$HOME/.config/eww/scripts/applaucher.sh"
+        # nwg-dock-hyprland -d -hd 60 -i 32 -w 5 -mb 10 -x -s themes/$DOCK_THEME/style.css -c "$HOME/.config/hypr/scripts/launcher.sh"
     else
         nwg-dock-hyprland -i 90 -w 5 -mb 2 -x -c "$HOME/.config/eww/scripts/applaucher.sh"
     fi