+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+Version 2.13.1
+--------------------------------------------------------
+- Add Dock autohide switch to sidebar.
+
 Version 2.13.0
 --------------------------------------------------------
 - Hyprland Configuration converted into new lua format.
@@ -5,6 +9,8 @@ Version 2.13.0
 - hyprshutdown added to Power App
 - Add .conf cleanup script ~/.config/ml4w/scripts/ml4w-remove-conf to remove all .conf files from ~/.config/hypr that has been replaced with .lua
 - Wallpaper folder can be set directly in the wallpaper widget
+- Prepared for HyprMod. Can be installed from the Welcome App. Shows additional launch buttons on Welcome App and Sidebar if installed.
+- Drop Shadows added to Quickshell PanelWindows for UI consistance.
 
 Version 2.12.3
 --------------------------------------------------------